# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BubblePilot API
  version: 2.0.0
  description: BubblePilot 管理会话、二次验证、消息归档、版本化工作流、AI Provider 路由、执行与审计接口。
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/memes:
    get:
      operationId: getapiv1memes
      description: 分页查询表情包；query、enabled、status、offset、limit
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    post:
      operationId: postapiv1memes
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, name]
              properties:
                file: { type: string, format: binary }
                name: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 2000 }
                tags:
                  { type: string, description: JSON 字符串数组，最多20个标签。 }
      description: 上传单张图片（multipart，file/name/description/tags），按内容哈希去重
      security:
        - sessionAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/memes/{id}:
    get:
      operationId: getapiv1memesid
      description: 读取素材元数据
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    put:
      operationId: putapiv1memesid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, name]
              properties:
                expectedVersion: { type: integer, minimum: 1 }
                name: { type: string, minLength: 1, maxLength: 120 }
                description: { type: string, maxLength: 2000, default: "" }
                tags:
                  type: array
                  maxItems: 20
                  items: { type: string, minLength: 1, maxLength: 40 }
                summary: { type: string, maxLength: 2400 }
                enabled: { type: boolean }
      description: 修改素材，expectedVersion 并发校验
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteapiv1memesid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer, minimum: 1 }
      description: 逻辑删除素材，expectedVersion 并发校验
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/memes/{id}/thumbnail:
    get:
      operationId: getapiv1memesidthumbnail
      description: 鉴权读取 PNG 首帧缩略图
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/memes/{id}/file:
    get:
      operationId: getapiv1memesidfile
      description: 鉴权读取原文件
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/memes/{id}/summary:
    post:
      operationId: postapiv1memesidsummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer, minimum: 1 }
                candidate: { type: boolean, default: true }
      description: 提交摘要任务，expectedVersion 必填；candidate 默认 true
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/memes/{id}/summary/adopt:
    post:
      operationId: postapiv1memesidsummaryadopt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer, minimum: 1 }
      description: 采用候选摘要，expectedVersion 必填
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 操作成功，返回 data；图片接口返回图片字节。
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionAuth:
      type: apiKey
      in: cookie
      name: bubblepilot_session
...
  responses:
    Error:
      description: 请求失败。
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message, correlationId]
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  correlationId:
                    type: string
                    format: uuid